feat: migration优化

+++ migrations/2020_05_29_000000_add_push_subscriptions_table.php
@@ -11,17 +11,24 @@
 
 use Flarum\Database\Migration;
 use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Schema\Builder;
 
-return Migration::createTable(
-    'push_subscriptions',
-    function (Blueprint $table) {
-        $table->increments('id');
-        $table->string('endpoint')->unique();
-        $table->string('vapid_public_key');
-        $table->string('keys')->nullable();
-        $table->timestamp('expires_at')->nullable();
-        $table->unsignedInteger('user_id');
+return [
+    'up' => function (Builder $schema) {
+        if (!$schema->hasTable('push_subscriptions')) {
+            $schema->create('push_subscriptions', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('endpoint')->unique();
+                $table->string('vapid_public_key');
+                $table->string('keys')->nullable();
+                $table->timestamp('expires_at')->nullable();
+                $table->unsignedInteger('user_id');
 
-        $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
+    },
+    'down' => function (Builder $schema) {
+        // 不执行删除操作
     }
-);
+];
