+++ migrations/2024_04_01_000000_add_firebase_push_subscriptions_table.php
@@ -11,14 +11,21 @@
 
 use Flarum\Database\Migration;
 use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Schema\Builder;
 
-return Migration::createTable(
-    'firebase_push_subscriptions',
-    function (Blueprint $table) {
-        $table->increments('id');
-        $table->string('token')->unique();
-        $table->unsignedInteger('user_id');
+return [
+    'up' => function (Builder $schema) {
+        if (!$schema->hasTable('firebase_push_subscriptions')) {
+            $schema->create('firebase_push_subscriptions', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('token')->unique();
+                $table->unsignedInteger('user_id');
 
-        $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
+    },
+    'down' => function (Builder $schema) {
+        // 不执行删除操作
     }
-);
+];
